some updates
git-svn-id: svn+ssh://combine-101/home/svn_bio/shrimp/trunk@700 450597fd-3c47-0410-84d8-f85f2794411d

+++ mergesam/sam2pretty_lib.h
@@ -23,11 +23,11 @@ struct pretty {
 	uint32_t* cigar_lengths;
 	//for colour space
 	bool colour_space;
-	size_t genome_start_padded;
-	size_t genome_end_padded;
-	size_t genome_start_unpadded;
-	size_t mate_genome_start_unpadded;
-	size_t genome_end_unpadded;
+	unsigned long genome_start_padded;
+	unsigned long genome_end_padded;
+	unsigned long genome_start_unpadded;
+	unsigned long mate_genome_start_unpadded;
+	unsigned long genome_end_unpadded;
 	int clipped_read_start;
 	int clipped_read_end;
 	int pretty_clipped_read_start;
